Add an example results chart
- Uses ApexCharts to create a simple stacked bar.
- Hardcodes the values based on slider inputs. This will be changed to
  use real API responses once implemented.

I wasn't able to find a way to nicely place the legend labels within the
bar itself, as most series start at zero (and so such labels would
overlap and be unreadable). I also can't find a way to offset the bar to
the left, which would allow us to add annotations for Russian gas and
"Compensation". It may be that we need to use D3 if we want that kind of
control.