Add basic HTTP requests to ETEngine
This uses thunks for updating inputs and presets and to trigger a
request to ETEngine. This feels clunky to me: it would be nicer to have
the input/preset change to be a normal state change with the API request
happening as a thunk.

However that doesn't really work since changes in state (input or preset
value) should not have side-effects. Still looking for a better way...