Store state in localStorage for faster reloads
Store the state of the inputs slice in localStorage whenever a request
completes successfully, allowing the page to be reloaded with the same
scenario active. This includes storing the slider values and gquery
results.

If the app is changed to add or remove sliders or gqueries, the stored
state is invalidated and a new scenario is created.