Add support for new thermostat input
The ETM now has a custom input for setting the heat demand reduction
caused by thermostats. The sliders have been adjusted.

Since scenarios created by this tool are no longer compatible with the
ETM front-end (on account of using a custom input), the ETM link now
goes to the home page rather than a copy of the scenario.